+++ R/gendistance.R
@@ -50,6 +50,9 @@
 #'trickle-in randomized trials.  When set, non-NA values from this column are
 #'replaced with the value 1.  This prevents records with previously assigned
 #'treatments (the \sQuote{reservior}) from matching each other.
+#'@param outRawDist a logical, indicating if the raw distance matrix should also
+#'be returned.  The raw form is before distance modifiers such as \sQuote{prevent}
+#'take effect.
 #'@param \dots Additional arguments, not used at this time.
 #'@return a list object with several elements
 #'
@@ -70,6 +73,8 @@
 #'  \item{missing.weight}{weight to apply to missingness indicator columns}
 #'
 #'  \item{ndiscard}{number of elements that will match phantoms}
+#'
+#'  \item{rawDist}{raw distance matrix, only provided if \sQuote{outRawDist} is TRUE}
 #'@exportMethod gendistance
 #'@author Cole Beck
 #'@seealso \code{\link{distancematrix}}
@@ -90,11 +95,11 @@
 setGeneric("gendistance", function(covariate, idcol=NULL, weights=NULL,
            prevent=NULL, force=NULL, rankcols=NULL, missing.weight=0.1,
            ndiscard=0, singular.method='solve', talisman=NULL,
-           prevent.res.match=NULL, ...) standardGeneric("gendistance"))
+           prevent.res.match=NULL, outRawDist = FALSE, ...) standardGeneric("gendistance"))
 setMethod("gendistance", "data.frame", function(covariate, idcol=NULL,
           weights=NULL, prevent=NULL, force=NULL, rankcols=NULL,
           missing.weight=0.1, ndiscard=0, singular.method='solve',
-          talisman=NULL, prevent.res.match=NULL, ...) {
+          talisman=NULL, prevent.res.match=NULL, outRawDist = FALSE, ...) {
     nr <- nrow(covariate)
     nc <- ncol(covariate)
     stopifnot(nr > 0)
@@ -280,6 +285,9 @@ setMethod("gendistance", "data.frame", function(covariate, idcol=NULL,
     maxval <- Inf
     # add back row names
     dimnames(mdists) <- list(myrownames, myrownames)
+    if(outRawDist) {
+        rawDist <- mdists
+    }
 
     # penalize "prevent" columns with matching values by setting distance to maxval
     # prevent is a vector of column names found in bad.data
@@ -323,6 +331,13 @@ setMethod("gendistance", "data.frame", function(covariate, idcol=NULL,
     diag(mdists) <- maxval
     # convert matrix to data frame
     mdists <- as.data.frame(mdists)
-
-    list(dist=mdists, cov=covariate, ignored=bad.data, weights=weights, prevent=prevent, mates=mateIDs, rankcols=rankcols, missing.weight=missing.weight, ndiscard=nphantoms)
+    res <- list(dist=mdists, cov=covariate, ignored=bad.data, weights=weights,
+        prevent=prevent, mates=mateIDs, rankcols=rankcols,
+        missing.weight=missing.weight, ndiscard=nphantoms
+    )
+    if(outRawDist) {
+        diag(rawDist) <- maxval
+        res[['rawDist']] <- as.data.frame(rawDist)
+    }
+    res
 })