+++ NEWS.md
@@ -1,5 +1,6 @@
 # nbpmatching 1.5.3
 * additional check for singular matrix in `gendistance`; thanks Jonathan Chipman
+* new `gendistance` argument `outRawDist`
 
 # nbpmatching 1.5.2
 * properly register Fortran calls